# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小泉动漫后端API接口 - 支持内容管理、评论、投票、用户认证、站外资源等",
        "title": "小泉动漫 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/admin/audit/{id}": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员审核内容（通过/拒绝）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "审核内容",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "内容ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "审核信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "admin_id": {
                                    "type": "integer"
                                },
                                "remark": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SwaggerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Content"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/admin/claims": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员查看内容认领申请列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "认领申请列表",
                "parameters": [
                    {
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "type": "string",
                        "description": "状态筛选",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "内容ID筛选",
                        "name": "content_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SwaggerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "list": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/models.Claim"
                                                    }
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/claims/{id}/handle": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员通过或拒绝内容认领申请",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "处理认领申请",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "认领申请ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "操作(approve/reject)和备注",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "action": {
                                    "type": "string"
                                },
                                "remark": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SwaggerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Claim"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/admin/comments/reports": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员查看评论举报列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "举报列表",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/admin/comments/reports/{id}/handle": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员处理评论举报，标记为已处理",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "处理举报",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "举报ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SwaggerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.CommentReport"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/admin/content/all": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员查看所有内容列表（支持筛选）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "所有内容",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "标签筛选",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "video",
                            "image",
                            "text",
                            "link"
                        ],
                        "type": "string",
                        "description": "内容类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "approved",
                            "pending",
                            "rejected"
                        ],
                        "type": "string",
                        "description": "审核状态",
                        "name": "audit_status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "关键词搜索",
                        "name": "keyword",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SwaggerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "list": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/models.Content"
                                                    }
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "page_size": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "total_page": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/content/{id}/author": {
            "put": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员将内容转移给其他用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "更新内容作者",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "内容ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "目标用户ID",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/admin/files/clean": {
            "delete": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员清理未被任何内容引用的上传文件和缩略图",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "清理孤立文件",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SwaggerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "deleted_count": {
                                                    "type": "integer"
                                                },
                                                "deleted_files": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    }
                }
            }
        },
        "/admin/pending": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "管理员查看待审核内容列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "待审核内容",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.SwaggerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "list": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/models.Content"
                                                    }
                                                },
                                                "page": {
                                                    "type": "integer"
                                                },
                                                "page_size": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "total_page": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/admin/users/{id}": {
            "delete": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "删除用户及其所有内容",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理"
                ],
                "summary": "删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.SwaggerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AuditStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "AuditStatusPending",
                "AuditStatusApproved",
                "AuditStatusRejected"
            ]
        },
...
        "models.Content": {
            "type": "object",
            "properties": {
                "audit_status": {
                    "$ref": "#/definitions/models.AuditStatus"
                },
                "big_tag_id": {
                    "type": "integer"
                },
                "compressed_path": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "file_path": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "platform": {
                    "type": "string"
                },
                "small_tag_id": {
                    "type": "integer"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumb_path": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.ContentType"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "user_id": {
                    "type": "integer"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
        "utils.SwaggerResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                }
            }
        }
    },
    "securityDefinitions": {
        "SessionAuth": {
            "description": "登录后自动携带的session cookie",
            "type": "apiKey",
            "name": "session_id",
            "in": "cookie"
        }
    }
}